+++ packages/@expo/cli/src/start/server/metro/inspector-proxy/index.ts
@@ -11,14 +11,20 @@ export { ExpoInspectorProxy } from './proxy';
 
 const debug = require('debug')('expo:metro:inspector-proxy') as typeof console.log;
 
-export function createInspectorProxy(metroBundler: MetroBundlerDevServer, projectRoot: string) {
+export function createInspectorProxy(
+  metroBundler: MetroBundlerDevServer,
+  projectRoot: string
+): ExpoInspectorProxy {
   debug('Expo inspector proxy enabled');
 
   // The device is slightly more complicated, we need to extend that class
   const ExpoInspectorDevice = createInspectorDeviceClass(metroBundler, Device);
 
   const inspectorProxy = new ExpoInspectorProxy(
-    new MetroInspectorProxy(projectRoot),
+    new MetroInspectorProxy(projectRoot, '', null, {
+      enableNewDebugger: true,
+      enableOpenDebuggerRedirect: true,
+    }),
     ExpoInspectorDevice
   );