+++ packages/@expo/cli/CHANGELOG.md
@@ -19,6 +19,7 @@
 ### 🐛 Bug fixes
 
 - Prevent log-spew when running prebuild in debug mode. ([#25434](https://github.com/expo/expo/pull/25434) by [@EvanBacon](https://github.com/EvanBacon))
+- Fix @react-native/dev-middleware types for react-native 0.73. ([#25513](https://github.com/expo/expo/pull/25513) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 
 ### 💡 Others